+++ modeltrans/__init__.py
@@ -1,3 +1,3 @@
 default_app_config = 'modeltrans.apps.RegistrationConfig'
 
-__version__ = '0.0.7'
+__version__ = '0.0.8'